move `assertEqualWithPrecision` to a separate file
in this way we can import it only where we need it
